+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/bitrise-core/bitrise-plugins-io/cmd"
+import "github.com/bitrise-io/bitrise-plugins-io/cmd"
 
 func main() {
 	cmd.Execute()